+++ Realm/RLMObjectSchema_Private.hpp
@@ -39,6 +39,7 @@
 
 // The Realm retains its object schemas, so they need to not retain the Realm
 @property (nonatomic, unsafe_unretained) RLMRealm *realm;
+@property (nonatomic, readwrite, strong) NSDictionary *defaultValues;
 @property (nonatomic, readonly) tightdb::Table *table;
 
 // returns a cached or new schema for a given object class